code refactor
base code now reside in src/service/geonode.js
