+++ src/services/geonode.js
@@ -1,7 +1,13 @@
 import MapConfigTransformService from 'boundless-sdk/services/MapConfigTransformService';
 import MapConfigService from 'boundless-sdk/services/MapConfigService';
-import {getCRSFToken, removeTrailingSlash} from '../helper';
-import {edit_map_endpoint, NEW_MAP_ENDPOINT} from '../constants/server'
+import {
+  getCRSFToken,
+  removeTrailingSlash
+} from '../helper';
+import {
+  edit_map_endpoint,
+  NEW_MAP_ENDPOINT
+} from '../constants/server'
 
 import 'whatwg-fetch';
 
@@ -23,7 +29,7 @@ const saveMethod = (id = undefined) => {
   return id ? 'PUT' : 'POST';
 };
 const checkStatus = (response) => {
-  if(response.status >= 200 && response.status < 300) {
+  if (response.status >= 200 && response.status < 300) {
     return response;
   } else {
     var error = new Error(response.statusText);
@@ -31,10 +37,10 @@ const checkStatus = (response) => {
     throw error;
   }
 };
-export const saveToGeonode = (server,config, id = undefined) => {
+export const saveToGeonode = (server, config, id = undefined) => {
   var request = createRequestObject(saveMethod(id), JSON.stringify(config));
   var requestPath = removeTrailingSlash(server) + saveEndPoint(id);
-  return fetch(requestPath,request)
+  return fetch(requestPath, request)
     .then(checkStatus)
     .then((response) => response.json())
     .catch((ex) => Promise.reject(ex));
@@ -56,4 +62,35 @@ export const getLocalGeoServer = (sources, baseUrl) => {
       return source;
     }
   }
+};
+export const createThumbnail = (obj_id, map) => {
+  map.once('postcompose', function(evt) {
+    var canvas = evt.context.canvas;
+    canvas.toBlob(function(blob) {
+      var url = window.location.pathname.replace('/view', '');
+
+      if (typeof obj_id != 'undefined' && url.indexOf('new')) {
+        url = url.replace('new', obj_id);
+      }
+      url += '/thumbnail/react';
+      var reader = new window.FileReader();
+      reader.readAsDataURL(blob);
+      reader.onloadend = function() {
+        fetch(url, {
+          method: 'POST',
+          credentials: "same-origin",
+          headers: {
+            'Accept': 'application/json',
+            'Content-Type': 'application/json',
+            "X-CSRFToken": getCRSFToken()
+          },
+          body: JSON.stringify({
+            image: reader.result
+          })
+        })
+      }
+
+    });
+  });
+  map.renderSync();
 };
